polished user input!!! FINALLY!
There will no longer be an error whenever a user puts in letters instead of numbers! If they were to input something other than a number, the program prints, "Don't be Silly! Enter a NUMBER for the __ Coordinate...", and proceeds to asking the user to input that specific coordinate again :)