simplified the program
i put the polished user input methods in a singular method and I also simplified the converting of the coordinates into one method. This also helps that little logic error where the Y coordinate was displayed as a double while the X and Y coordinates displayed as an integer. Now, they are all shown as integers (whole numbers) 🔢